Improved CSS for the max width, including w/ <pre>
Instead of having the <pre> have a fixed width, I prefer to have them
respect their parent's – but then said parents need a fixed width. Which
can be done while respecting the header layout in small viewports,  at
the cost of a extra container <div> – something I can live with. And it
simplifies the CSS a bit.