Series
This one was difficult, because it dived into Zenweb to understand how
the built-in mechanism for series work. Once you get it, it's clear,
but boy did it take me time to get it.

Now that the generation is aware of the series and theirs pages, the
layouts have been simplified a bit. Plus, some CSS for the design and
some cleanup in the markdown files to fix rendering issues.